pyproject.toml
[tool.poetry]
name = "biomgi_publications"
version = "0.1.0"
description = ""
authors = ["Rainforest Cheng <r08521610@ntu.edu.tw>"]

[tool.poetry.dependencies]
python = "^3.8"
mdutils = "^1.3.0"
bibtexparser = "^1.2.0"
python-frontmatter = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
